+++ HISTORY.md
@@ -1,4 +1,8 @@
 #### 1.0.2
-
   - **is_instance_of** got alias **is_a**
   - **is_not_instance_of** got alias **is_not_a**
+
+
+#### 1.0.3
+  - **is_instance_of** renamed to **is_kind_of**
+  - **is_not_instance_of** renamed to **is_not_kind_of**